Fix TPay Issue

+++ PaymentDomain/Features/TPay/DTO/Command/TPayEndpointInitializerRequest.cs
@@ -3,7 +3,7 @@
     public class TPayEndpointInitializerRequest
     {
         public string ProductName { get; set; }
-        public decimal ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
         public string OperatorCode { get; set; }
         public string MSISDN { get; set; }
         public int Language { get; set; }
